refactor(colorContext):merge context and provider, then move it to context folder
