--- src/colorContext/colorContext.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const ColorContext = React.createContext({});
-
-export default ColorContext;
